+++ src/main/java/likelion/festival/booth/ui/BoothController.java
@@ -6,10 +6,12 @@ import likelion.festival.booth.application.dto.BoothFilterDto;
 import likelion.festival.dto.*;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.booth.application.BoothService;
+import likelion.festival.menu.application.dto.MenuRequestDto;
+import likelion.festival.menu.application.dto.MenuResponseDto;
 import likelion.festival.service.CommentService;
 import likelion.festival.service.ImageService;
 import likelion.festival.service.LikesService;
-import likelion.festival.service.MenuService;
+import likelion.festival.menu.application.MenuService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
